Moved MenuExt to androidshared
